pyproject.toml
[project]
name = "rl-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gymnasium[accept-rom-license,atari]>=1.2.1",
    "opencv-python>=4.12.0.88",
    "rl-zoo3>=2.7.0",
    "tqdm>=4.67.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "scikit-learn>=1.7.2",
    "h5py>=3.15.1",
    "pillow>=12.0.0",
    "cupy-cuda12x>=13.6.0",
]

[[tool.uv.index]]
url="https://pypi.tuna.tsinghua.edu.cn/simple"
default=true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true


[tool.uv.sources]
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
